feat: add draft datasources

+++ src/lib/pages/datasource/update-datasource.ts
@@ -1,5 +1,5 @@
 import {Component, Injector, OnInit} from '@angular/core';
-import {ActivatedRoute} from '@angular/router';
+import {ActivatedRoute, Router} from '@angular/router';
 import {DatePipe} from '@angular/common';
 import {AuthenticationService} from '../../services/authentication.service';
 import {Subscription} from 'rxjs';
@@ -9,6 +9,7 @@ import {DatasourceFormComponent} from "./datasource-form.component";
 import {FormControlService} from "../../../dynamic-catalogue/services/form-control.service";
 import {ConfigService} from "../../services/config.service";
 import {DatasourceService} from "../../services/datasource.service";
+import {pidHandler} from "../../shared/pid-handler/pid-handler.service";
 
 @Component({
     selector: 'app-update-datasource',
@@ -19,14 +20,16 @@ export class UpdateDatasource extends DatasourceFormComponent implements OnInit 
   private sub: Subscription;
 
   constructor(public route: ActivatedRoute,
+              public router: Router,
               public authenticationService: AuthenticationService,
               protected datasourceService: DatasourceService,
               protected injector: Injector,
               public datePipe: DatePipe,
               public navigator: NavigationService,
               public dynamicFormService: FormControlService,
-              public config: ConfigService) {
-    super(injector, authenticationService, datasourceService, route, dynamicFormService, config);
+              public config: ConfigService,
+              public pidHandler: pidHandler) {
+    super(injector, authenticationService, datasourceService, route, router, dynamicFormService, config, pidHandler);
     this.editMode = true;
   }
 
@@ -43,18 +46,22 @@ export class UpdateDatasource extends DatasourceFormComponent implements OnInit 
         // this.datasourceId = params['datasourceId'];
         this.datasourceId = this.route.snapshot.paramMap.get('datasourceId');
         const pathName = window.location.pathname;
-        if (pathName.includes('draft-datasource/update')) {
-          this.pendingResource = true;
-        }
-        // this.datasourceService.getService(this.resourceId).subscribe(service => {
-        // this.datasourceService[this.pendingResource ? 'getPendingDatasource' : 'getDatasourceBundleById'](this.datasourceId, this.catalogueId)
-        this.datasourceService.getDatasourceBundleById(this.datasourceId, this.catalogueId)
-          .subscribe(dsBundle => {
+        if (pathName.includes('draft-datasource/update')) this.pendingResource = true;
+        if (this.pendingResource) {
+          this.datasourceService.getDraftDatasource(this.datasourceId).subscribe(ds => {
+                this.payloadAnswer = {'answer': {datasource: ds}};
+              },
+              err => this.errorMessage = 'Could not get the data for the requested datasource. ' + err.error
+            );
+        } else {
+          // this.datasourceService[this.pendingResource ? 'getPendingDatasource' : 'getDatasourceBundleById'](this.datasourceId, this.catalogueId)
+          this.datasourceService.getDatasourceBundleById(this.datasourceId, this.catalogueId).subscribe(dsBundle => {
               this.payloadAnswer = {'answer': {datasource: dsBundle.datasource}};
               ResourceService.removeNulls(dsBundle.datasource);
             },
-            err => this.errorMessage = 'Could not get the data for the requested service. ' + err.error
+            err => this.errorMessage = 'Could not get the data for the requested datasource. ' + err.error
           );
+        }
       });
     }
   }
